chore: jwt 인증이 되었을때만 마이페이지 접근되도록 swaggerConfig 수정 및 관련 조건 수정 & refactor: 인증 적용을 위한 userController 분리
