+++ src/main/java/com/travelservice/config/SwaggerConfig.java
@@ -3,17 +3,31 @@ package com.travelservice.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityScheme;
 
 @Configuration
 public class SwaggerConfig {
 	@Bean
 	public OpenAPI openApi() {
+		final String securitySchemeName = "bearerAuth";
+
 		return new OpenAPI()
 			.info(new Info()
 				.title("API 문서 제목")
 				.version("v1.0")
-				.description("API 설명"));
+				.description("API 설명"))
+			.components(new Components()
+				.addSecuritySchemes(securitySchemeName,
+					new SecurityScheme()
+						.name(securitySchemeName)
+						.type(SecurityScheme.Type.HTTP)
+						.scheme("bearer")
+						.bearerFormat("JWT")
+				)
+			);
 	}
+
 }
